New position

+++ CPP_functions/NewPosition/NewPositions.h
@@ -1,8 +1,11 @@
 #ifndef NEWPOSITION_NEWPOSITIONS_H
 #define NEWPOSITION_NEWPOSITIONS_H
+#include <vector>
+#include "math.h"
+#include <utility>
 
 std::pair<int, int>
-NewPosition(std::pair<int, int> cur_pos, const std::vector<std::pair<int, int>> &neighbours, int cool, int C);
+NewPosition(std::pair<int, int> cur_pos, const std::vector<std::pair<int, int>> &neighbours, double cool, double C);
 
 
 #endif //NEWPOSITION_NEWPOSITIONS_H
